Added in a cast to cast the pointer value (end - s) into an integer.
git-svn-id: http://llvm.org/svn/llvm-project/safecode/trunk@113203 91177308-0d34-0410-b5e6-96231b3b80d8

+++ runtime/DebugRuntime/CStdLib.cpp
@@ -34,7 +34,7 @@ extern "C" {
 
   size_t strnlen_opt(const char *s, size_t maxlen) {
     const char *end = (const char *)memchr(s, '\0', maxlen);
-    return (end ? end - s : maxlen);
+    return (end ? ((size_t) (end - s)) : maxlen);
   }
 }
